Adding Genetic Algorithm code.
Adds Genetic Algorithm code to the repository. GA was developed as a means to find the best set of parameters for the JRip algorithm executed in the software Weka.
Also adds a better separation between the algorithms, by creating a directory for each one.
Finally, adds a git commit template for safe keeping.

Commit:Add

TaskNumber: